Rename EntitiesDescriptor to MetadataInstance
Part of a series of changes that restructures the top level of this data
model into two pieces representing the input and output of the
saml-service. EntitiesDescriptor is currently something of a generic
container and has no necessary relationship to the resulting
EntitiesDescriptor in the output.
